Add unit test for Invoking Contract by Contract
